+++ libs/rudder-integration-appsflyer-react-native/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.5.8](https://github.com/rudderlabs/rudder-sdk-react-native/compare/rudder-integration-appsflyer-react-native@1.5.7...rudder-integration-appsflyer-react-native@1.5.8) (2023-08-09)
+
+
+### Bug Fixes
+
+* appsflyer native event emitter pop up issue ([#228](https://github.com/rudderlabs/rudder-sdk-react-native/issues/228)) ([1b911b1](https://github.com/rudderlabs/rudder-sdk-react-native/commit/1b911b1b66ed76c3c53d9fd8638d755d9b8a3113))
+
 ## [1.5.7](https://github.com/rudderlabs/rudder-sdk-react-native/compare/rudder-integration-appsflyer-react-native@1.5.6...rudder-integration-appsflyer-react-native@1.5.7) (2023-06-22)